feat: ArgumentResolver 추가

+++ src/main/java/cart/service/CartDeleteService.java
@@ -1,5 +1,9 @@
 package cart.service;
 
+import java.util.Optional;
+
+import cart.domain.user.User;
+import cart.entiy.CartEntity;
 import cart.repository.CartRepository;
 import org.springframework.stereotype.Service;
 
@@ -12,7 +16,13 @@ public class CartDeleteService {
         this.cartRepository = cartRepository;
     }
 
-    public void delete(final long id) {
-        cartRepository.deleteById(id);
+    public void delete(final long id, final User user) {
+        final Optional<CartEntity> cart = cartRepository.findById(id);
+        final String email = cart.orElseThrow(CartNotFoundException::new).getEmail();
+        if (user.getEmail().getValue().equals(email)) {
+            cartRepository.deleteById(id);
+            return;
+        }
+        throw new CartOwnerMismatchException();
     }
 }
